requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Slack-Events==0.0.2
h11==0.9.0
httptools==0.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
mangum==0.9.2
MarkupSafe==1.1.1
multidict==4.7.6
pyee==7.0.2
requests==2.23.0
six==1.15.0
slackclient==2.6.1
slackeventsapi==2.1.0
typing-extensions==3.7.4.2
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
yarl==1.4.2
